fix: reduce neuroskill sensor complexity
