+++ packages/lib/sensor-neuroskill/src/neuroskill-sensor.ts
@@ -1,4 +1,5 @@
 import type { SignalSource, UserSignal } from "@koi/core";
+import { estimateCognitiveState, extractFeatures, validateFrame } from "./signal-processing.js";
 
 export type NeuroSignalKind = "eeg" | "emg" | "eog" | "ecg" | "unknown";
 export type NeuroSkillConnectionState = "idle" | "connected" | "disconnected" | "error";
@@ -169,22 +170,6 @@ function resolveConfig(config: NeuroSkillSensorConfig): ResolvedConfig {
   };
 }
 
-function roundMetric(value: number): number {
-  if (!Number.isFinite(value)) return 0;
-  return Math.round(value * 1000) / 1000;
-}
-
-function clamp01(value: number): number {
-  if (!Number.isFinite(value)) return 0;
-  if (value < 0) return 0;
-  if (value > 1) return 1;
-  return roundMetric(value);
-}
-
-function isRecord(value: unknown): value is Readonly<Record<string, unknown>> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
 function parseJson(data: unknown): unknown {
   if (typeof data !== "string") return data;
   try {
@@ -193,169 +178,6 @@ function parseJson(data: unknown): unknown {
   } catch {
     return null;
   }
-}
-
-function validateFrame(input: unknown, maxSamples: number): NeuroSignalFrame | null {
-  if (!isRecord(input)) return null;
-  const timestamp = input.timestamp;
-  const samplingRateHz = input.samplingRateHz;
-  const channels = input.channels;
-  if (typeof timestamp !== "number" || typeof samplingRateHz !== "number") return null;
-  if (!Number.isFinite(timestamp) || !Number.isFinite(samplingRateHz)) return null;
-  if (samplingRateHz <= 0 || !Array.isArray(channels) || channels.length === 0) return null;
-
-  const parsedChannels = channels.flatMap((channel): readonly NeuroSignalChannel[] => {
-    if (!isRecord(channel)) return [];
-    const name = channel.name;
-    const rawSamples = channel.samples;
-    if (typeof name !== "string" || name.length === 0 || !Array.isArray(rawSamples)) return [];
-    if (!rawSamples.every((sample) => typeof sample === "number" && Number.isFinite(sample))) {
-      return [];
-    }
-    const samples = rawSamples.slice(0, maxSamples);
-    if (samples.length === 0) return [];
-    const type = typeof channel.type === "string" ? channel.type : "unknown";
-    return [{ name, type: normalizeKind(type), samples }];
-  });
-
-  if (parsedChannels.length === 0) return null;
-  return { timestamp, samplingRateHz, channels: parsedChannels };
-}
-
-function normalizeKind(kind: string): NeuroSignalKind {
-  if (kind === "eeg" || kind === "emg" || kind === "eog" || kind === "ecg") return kind;
-  return "unknown";
-}
-
-function mean(samples: readonly number[]): number {
-  if (samples.length === 0) return 0;
-  return samples.reduce((sum, sample) => sum + sample, 0) / samples.length;
-}
-
-function median(samples: readonly number[]): number {
-  if (samples.length === 0) return 0;
-  const sorted = [...samples].sort((a, b) => a - b);
-  const midpoint = Math.floor(sorted.length / 2);
-  if (sorted.length % 2 === 1) return sorted[midpoint] ?? 0;
-  return ((sorted[midpoint - 1] ?? 0) + (sorted[midpoint] ?? 0)) / 2;
-}
-
-function variance(samples: readonly number[], sampleMean: number): number {
-  if (samples.length === 0) return 0;
-  return samples.reduce((sum, sample) => sum + (sample - sampleMean) ** 2, 0) / samples.length;
-}
-
-function rms(samples: readonly number[]): number {
-  if (samples.length === 0) return 0;
-  return Math.sqrt(samples.reduce((sum, sample) => sum + sample ** 2, 0) / samples.length);
-}
-
-function removeNoise(
-  samples: readonly number[],
-  sigma: number,
-): { readonly samples: readonly number[]; readonly rejected: number } {
-  const baseline = median(samples);
-  const centered = samples.map((sample) => sample - baseline);
-  const absoluteDeviations = centered.map((sample) => Math.abs(sample));
-  const robustStd = median(absoluteDeviations) * 1.4826;
-  const limit = robustStd > 0 ? robustStd * sigma : sigma;
-  const rejected = centered.filter((sample) => Math.abs(sample) > limit).length;
-  return {
-    samples: centered.map((sample) => Math.max(-limit, Math.min(limit, sample))),
-    rejected,
-  };
-}
-
-function zeroCrossingRate(samples: readonly number[]): number {
-  if (samples.length < 2) return 0;
-  const crossings = samples.slice(1).filter((sample, index) => {
-    const previous = samples[index] ?? 0;
-    return (previous < 0 && sample >= 0) || (previous >= 0 && sample < 0);
-  }).length;
-  return roundMetric(crossings / (samples.length - 1));
-}
-
-function computeBandPowers(samples: readonly number[], samplingRateHz: number): NeuroBandPowers {
-  const bins = samples.map((_, index) => index).slice(1, Math.floor(samples.length / 2) + 1);
-  const powers = bins.map((bin) => {
-    const frequency = (bin * samplingRateHz) / samples.length;
-    const real = samples.reduce(
-      (sum, sample, index) => sum + sample * Math.cos((2 * Math.PI * bin * index) / samples.length),
-      0,
-    );
-    const imaginary = samples.reduce(
-      (sum, sample, index) => sum - sample * Math.sin((2 * Math.PI * bin * index) / samples.length),
-      0,
-    );
-    return { frequency, power: real ** 2 + imaginary ** 2 };
-  });
-
-  const sumBand = (low: number, high: number): number =>
-    roundMetric(
-      powers
-        .filter((entry) => entry.frequency >= low && entry.frequency < high)
-        .reduce((sum, entry) => sum + entry.power, 0),
-    );
-
-  return {
-    delta: sumBand(0.5, 4),
-    theta: sumBand(4, 8),
-    alpha: sumBand(8, 13),
-    beta: sumBand(13, 30),
-    gamma: sumBand(30, 100),
-  };
-}
-
-function extractFeatures(frame: NeuroSignalFrame, cfg: ResolvedConfig): NeuroSignalFeatures {
-  return {
-    timestamp: frame.timestamp,
-    samplingRateHz: frame.samplingRateHz,
-    channels: frame.channels.map((channel) => {
-      const filtered = removeNoise(channel.samples, cfg.noiseSigma);
-      const sampleMean = mean(filtered.samples);
-      const sampleVariance = variance(filtered.samples, sampleMean);
-      const peakAmplitude = filtered.samples.reduce(
-        (peak, sample) => Math.max(peak, Math.abs(sample)),
-        0,
-      );
-      return {
-        name: channel.name,
-        type: channel.type ?? "unknown",
-        sampleCount: filtered.samples.length,
-        mean: roundMetric(sampleMean),
-        rms: roundMetric(rms(filtered.samples)),
-        variance: roundMetric(sampleVariance),
-        peakAmplitude: roundMetric(peakAmplitude),
-        zeroCrossingRate: zeroCrossingRate(filtered.samples),
-        noiseRejectedSamples: filtered.rejected,
-        bandPowers: computeBandPowers(filtered.samples, frame.samplingRateHz),
-      };
-    }),
-  };
-}
-
-function aggregateBand(features: NeuroSignalFeatures, band: keyof NeuroBandPowers): number {
-  const eeg = features.channels.filter((channel) => channel.type === "eeg");
-  if (eeg.length === 0) return 0;
-  return eeg.reduce((sum, channel) => sum + channel.bandPowers[band], 0) / eeg.length;
-}
-
-function estimateCognitiveState(
-  features: NeuroSignalFeatures,
-  sampledAt: number,
-): CognitiveStateEstimate {
-  const theta = aggregateBand(features, "theta");
-  const alpha = aggregateBand(features, "alpha");
-  const beta = aggregateBand(features, "beta");
-  const gamma = aggregateBand(features, "gamma");
-  const total = theta + alpha + beta + gamma;
-  const confidence = clamp01(features.channels.length / 4);
-  if (total <= 0) return { ...EMPTY_STATE, sampledAt };
-
-  const attention = clamp01((beta + gamma) / total);
-  const fatigue = clamp01((theta + alpha * 0.5) / total);
-  const engagement = clamp01((beta + alpha) / total);
-  return { attention, fatigue, engagement, confidence, sampledAt };
 }
 
 function pruneEvents(
@@ -408,11 +230,8 @@ function defaultSocketFactory(url: string): NeuroSkillSocket {
   return wrapSocket(Reflect.construct(maybeWebSocket, [url]));
 }
 
-export function createNeuroSkillSensor(config: NeuroSkillSensorConfig = {}): NeuroSkillSensor {
-  const cfg = resolveConfig(config);
-  const subscribers = new Set<NeuroSkillSubscriber>();
-  // Mutable state is held inside the sensor instance; public snapshots are immutable copies.
-  const state: MutableState = {
+function createInitialState(): MutableState {
+  return {
     connectionState: "idle",
     socket: undefined,
     events: [],
@@ -420,15 +239,21 @@ export function createNeuroSkillSensor(config: NeuroSkillSensorConfig = {}): Neu
     latestFeatures: EMPTY_FEATURES,
     cognitiveState: EMPTY_STATE,
   };
+}
 
-  const rejectFrame = (): false => {
-    state.rejectedFrames += 1;
-    return false;
-  };
+function rejectFrame(state: MutableState): false {
+  state.rejectedFrames += 1;
+  return false;
+}
 
-  const record = (frameInput: unknown): boolean => {
+function createRecorder(
+  state: MutableState,
+  cfg: ResolvedConfig,
+  subscribers: ReadonlySet<NeuroSkillSubscriber>,
+): (frame: unknown) => boolean {
+  return (frameInput: unknown): boolean => {
     const frame = validateFrame(frameInput, cfg.maxSamplesPerChannel);
-    if (frame === null) return rejectFrame();
+    if (frame === null) return rejectFrame(state);
     const features = extractFeatures(frame, cfg);
     const cognitiveState = estimateCognitiveState(features, cfg.now());
     const event: NeuroSkillStateEvent = {
@@ -443,8 +268,10 @@ export function createNeuroSkillSensor(config: NeuroSkillSensorConfig = {}): Neu
     notifySubscribers(subscribers, event);
     return true;
   };
+}
 
-  const snapshot = (): NeuroSkillSnapshot => {
+function createSnapshotReader(state: MutableState, cfg: ResolvedConfig): () => NeuroSkillSnapshot {
+  return () => {
     const now = cfg.now();
     state.events = pruneEvents(state.events, cfg, now);
     return {
@@ -461,6 +288,41 @@ export function createNeuroSkillSensor(config: NeuroSkillSensorConfig = {}): Neu
       windowMs: cfg.windowMs,
     };
   };
+}
+
+function attachSocketHandlers(
+  socket: NeuroSkillSocket,
+  state: MutableState,
+  record: (frame: unknown) => boolean,
+): void {
+  const onMessage = (event: { readonly data?: unknown }): void => {
+    const parsed = parseJson(event.data);
+    if (parsed === null) {
+      rejectFrame(state);
+      return;
+    }
+    record(parsed);
+  };
+  const onClose = (): void => {
+    state.connectionState = "disconnected";
+  };
+  const onError = (): void => {
+    state.connectionState = "error";
+  };
+  socket.addEventListener?.("message", onMessage);
+  socket.addEventListener?.("close", onClose);
+  socket.addEventListener?.("error", onError);
+  socket.onmessage = onMessage;
+  socket.onclose = onClose;
+  socket.onerror = onError;
+}
+
+export function createNeuroSkillSensor(config: NeuroSkillSensorConfig = {}): NeuroSkillSensor {
+  const cfg = resolveConfig(config);
+  const subscribers = new Set<NeuroSkillSubscriber>();
+  const state = createInitialState();
+  const record = createRecorder(state, cfg, subscribers);
+  const snapshot = createSnapshotReader(state, cfg);
 
   return {
     name: cfg.name,
@@ -469,28 +331,7 @@ export function createNeuroSkillSensor(config: NeuroSkillSensorConfig = {}): Neu
       const socket = (cfg.socketFactory ?? defaultSocketFactory)(url);
       state.socket = socket;
       state.connectionState = "connected";
-      const onMessage = (event: { readonly data?: unknown }): void => {
-        const parsed = parseJson(event.data);
-        if (parsed === null) {
-          rejectFrame();
-          return;
-        }
-        record(parsed);
-      };
-      const onClose = (): void => {
-        state.connectionState = "disconnected";
-      };
-      const onError = (): void => {
-        state.connectionState = "error";
-      };
-      if (socket.addEventListener !== undefined) {
-        socket.addEventListener("message", onMessage);
-        socket.addEventListener("close", onClose);
-        socket.addEventListener("error", onError);
-      }
-      socket.onmessage = onMessage;
-      socket.onclose = onClose;
-      socket.onerror = onError;
+      attachSocketHandlers(socket, state, record);
     },
     disconnect() {
       state.socket?.close?.();
